Simplify worfkow
In resstock-euss.2022.1/resources/hpxml-measures/workflow remove uncessary files as tested by running the workflow with progressively fewer files. This is intended to reduce the runtime by having the python script not have to copy as much data between the base workflow folder into the working copy of the workflow folder which is done for each building file.